+++ Procfile.txt
@@ -1,1 +1,1 @@
-web: gunicorn app:api
+web: gunicorn api/app:app
