--- api/extensions.py
@@ -1,3 +1,0 @@
-from flask_pymongo import PyMongo
-
-mongo = PyMongo()
